requirements.txt
future
emoji
lxml
beautifulsoup4
requests
sqlalchemy==1.3.23 
python-telegram-bot==12.8
psycopg2-binary
feedparser
pynewtonmath
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
googletrans==3.1.0a0
#google_trans_new
jikanpy
speedtest-cli
coffeehouse
regex
bleach
git+https://github.com/starry69/python-markdown2.git
wikipedia
telethon==1.16.4
spamwatch
alphabet_detector
pyrate-limiter
cachetools
nekos.py
saucenao_api 
random2
opencv-python-headless 
telegraph 
redis
